% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above-grob.R
\name{above.grob}
\alias{above.grob}
\title{Superimpose two grobs, one above the other.}
\usage{
\method{above}{grob}(p, q, m = 1, n = 1)
}
\arguments{
\item{p}{A grob}

\item{q}{Another grob}

\item{m}{An integer}

\item{n}{Another integer}
}
\value{
A new grob obtained by superimposing p and q, one above the other.
}
\description{
Superimpose two grobs, one above the other.
}
