% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above.R
\name{above}
\alias{above}
\title{Superimpose two pictures, one above the other.}
\usage{
above(picture, ...)
}
\arguments{
\item{p}{A picture}
}
\value{
A picture
}
\description{
Superimpose two pictures, one above the other.
}

