% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beside.R
\name{beside}
\alias{beside}
\title{Superimpose two grobs, side-by-side.}
\usage{
beside(picture, ...)
}
\arguments{
\item{p}{A picture}
}
\value{
A picture
}
\description{
Superimpose two grobs, side-by-side.
}

