% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip a picture}
\usage{
flip(g)
}
\arguments{
\item{g}{A `grid::grob` picture}
}
\value{
A flipped `grid::grob` picture
}
\description{
`flip` returns a new picture made by flipping the input picture along
its vertical centre axis.
}
