% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonet.R
\name{nonet}
\alias{nonet}
\title{Nonet}
\usage{
nonet(a, s, d, f, g, h, j, k, l)
}
\arguments{
\item{a}{A picture}

\item{s}{Another picture}

\item{d}{Yet another picture}

\item{f}{The last picture}

\item{g}{A picture}

\item{h}{Another picture}

\item{j}{Yet another picture}

\item{k}{The last picture}

\item{l}{The last picture}
}
\value{
A grob
}
\description{
Nonet
}

