% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-picture.R
\name{plot.picture}
\alias{plot.picture}
\title{plot.picture}
\usage{
\method{plot}{picture}(x, ...)
}
\arguments{
\item{x}{A picture}

\item{...}{Anything else.}
}
\value{
grid.draw(picture)
}
\description{
plot.picture
}
