% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.R
\name{rot}
\alias{rot}
\title{Rotate a picture}
\usage{
rot(g, angle = 90)
}
\arguments{
\item{g}{A `grid::grob` picture}

\item{angle}{An angle}
}
\value{
A rotated `grid::grob` picture
}
\description{
`rot` returns a new picture, made by rotating the input picture. If no angle
is specified then the new picture will be rotated clockwise by 90 degrees. A
different angle can be specified with the `angle` argument.
}
