% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above.R
\name{above}
\alias{above}
\title{Superimpose two pictures, one above the other.}
\usage{
above(p, q, m, n)
}
\arguments{
\item{p}{A picture}

\item{q}{Another picture}

\item{m}{An integer}

\item{n}{Another integer}
}
\value{
A picture
}
\description{
Superimpose two pictures, one above the other.
}
