% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beside-grob.R
\name{beside.grob}
\alias{beside.grob}
\title{Superimpose two grobs side-by-side.}
\usage{
\method{beside}{grob}(p, q, m = 1, n = 1)
}
\arguments{
\item{g}{grob}
}
\value{
A grob
}
\description{
Superimpose two grobs side-by-side.
}

