% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beside.R
\name{beside}
\alias{beside}
\title{Superimpose two grobs, side-by-side.}
\usage{
beside(p, q, m = 1, n = 1)
}
\arguments{
\item{p}{grob}

\item{q}{grob}

\item{m}{An integer}

\item{n}{Another integer}
}
\value{
A grob

A picture
}
\description{
Superimpose two grobs, side-by-side.
}
